Retrive if a taxon is omnivore

Retrive if a taxon is omnivore or not, based on average prey trophic level or shortest trophic level method.

which_omnivore(graph, method = c("shortest", "average_prey"))

Arguments

graph
an igraph object.

method
computation method for the trophic level. Must be one of `shortest` or `average_prey`. See short_tl and prey_avg_tl for more informations.

Value

dataframe. First column containing the taxa and second if the taxon is omnivore or not.

References

Baiser, B., Gotelli, N. J., Buckley, H. L., Miller, T. E., & Ellison, A. M. (2012). Geographic stats::variation in network structure of a nearctic aquatic food web. Global Ecology and Biogeography.

Examples

data(aleutian)

which_omnivore(aleutian)

#> Methode: "shortest trophic level"

#>                 Predator Omnivore
#> 1                Mussels    FALSE
#> 2                Chitons    FALSE
#> 3             Gastropods    FALSE
#> 4                Limpets    FALSE
#> 5            Sea urchins    FALSE
#> 6              Asteroids    FALSE
#> 7               Decapods     TRUE
#> 8       Nearshore fishes    FALSE
#> 9           Harbor seals    FALSE
#> 10            Sea otters     TRUE
#> 11 Prehistoric aleut man     TRUE
#> 12  Epibenthic crustacea    FALSE